Rename action popup and simplify payment interaction store
