+++ frontend/src/payments/components/Experience.vue
@@ -2,7 +2,7 @@
 import { computed } from 'vue'
 import { storeToRefs } from 'pinia'
 
-import ActionPopup from '@/shared/components/ActionPopup.vue'
+import AppPopup from '@/shared/components/AppPopup.vue'
 import LoadingOverlay from '@/shared/components/LoadingOverlay.vue'
 import CurrencySelectorDialog from '@/payments/components/CurrencySelectorDialog.vue'
 import Section from '@/payments/components/Section.vue'
@@ -76,7 +76,7 @@ const onCloseTransferDialog = () => {
       @select="onCurrencySelect"
       @close="onCloseCurrencySelector"
     />
-    <ActionPopup
+    <AppPopup
       v-if="popupContent"
       :visible="isPopupVisible"
       :title="popupContent.title"
